requirements.txt
Flask
Redis
tensorflow
numpy
time
Pillow